gen-syscalls: generate an ASM table for the OS code
Instead of generating a C++ file, generate a plain ASM file.  By doing
this, we can avoid all of the pain that comes with C++ and C, and simply
assemble the table of symbols.

The most important thing this gets us is we can use 16-bit pointers, due
to the fact that we have <64k of flash.  This cuts our table size in half.

Signed-off-by: Sean Cross <xobs@kosagi.com>