common: KL02Z32: add section for flasheeprom
We emulate the Flash EEPROM, which eats up 1024 bytes of program space.
Add a section to the linker script that reflects this, and aligns it
to a 1024-byte boundary.

If the flash is not used, this section will not be included.

Note that it remains to be seen whether this will retain data across flashes.

Signed-off-by: Sean Cross <xobs@kosagi.com>