usbphy: fix threading, improve latency
This patch includes a bunch of things that probably should be separate.

First, it performs a syslock/sysunlock before waking the USB thread.

Next, the USB thread doesn't just sleep(), it does a sleepTimeout().
This is because sometimes we miss waking up the thread.

We also now precompute the address of the output buffer, which saves
about 16 cycles on ISR entry.  Every little bit helps.

Finally, some functions were marked static, and their names changed to
reflect the new static scope.

Signed-off-by: Sean Cross <sean@xobs.io>