USBCore: fix overflow bug on very large strings
The Arduino code uses a uint8_t for the string max length.  Unfortunately,
it has been observed to be as 1026 bytes coming from Windows.
Previously, this would get truncated to an 8-bit value, resulting in a
3-byte string getting sent out.

Use a uint32_t instead, to avoid the overflow and display the maximum
possible string length.

Signed-off-by: Sean Cross <sean@xobs.io>