support: Arduino: move io/time functions into "extern C"
These functions were converted from C++ to C, to make linking easier.
This was done in the previous few patches.  However, the function
prototypes were not marked as 'extern "C"'.  This meant the C++ compiler
couldn't find the symbols.

Move the definition of these functions into an 'extern "C"' section of
Arduino.h, in order to allow the C++ compiler to find them.

Signed-off-by: Sean Cross <sean@xobs.io>