support: syscalls: place all calls in the .text section
We want to compile with -flto, but if we do that the compiler will strip
away important functions such as __aeabi_idiv(), which is called every time
a divide is inserted.

Ideally we'd compile syscalls-app.cpp without -flto, but the Arduino build
system doesn't allow for doing that.  So we need to find another way around
it.

LTO works by placing each function in its own section.  We can manually
place every function in the .text section.  This will exempt it from lto,
while still allow it to remain unlinked unless it's used.

Signed-off-by: Sean Cross <sean@xobs.io>