support: begin moving syscalls into asm file
If we enable LTO, the most basic syscalls stop working.  This manifests
itself in a failure to compile when you have a division operation.

If we disable LTO, most things work fine, but the resulting binary ends
up larger than it ought to be.

The correct solution would be to disable LTO for only the
syscalls-app.cpp file.  However, Arduino doesn't allow for per-file
rules.

As a complete hack, generate an ASM file that looks similar enough to
the ASM file generated by compiling syscalls-app.cpp.

This results in an assembly file that links properly without LTO.  As a
bonus, we still get LTO on the higher-level C++ functions.

Signed-off-by: Sean Cross <sean@xobs.io>