support: Serial: wrap object in macro for dead-code removal
Replace the Serial object with a macro that calls a function named
serialWrapper().  This function simply returns a statically-allocated
HardwareSerial object.

We do this so that the HardwareSerial object will be garbage-collected
and removed if there are no Serial calls in the program.  This saves
250 bytes from programs that do not use Serial, and more importantly
is compatible with gcc 4.8, as used by Arduino.

Signed-off-by: Sean Cross <sean@xobs.io>