projects: palette-template: convert to template
Convert from the text-editor format to something that can be patched by
search-and-replace.

This is because we will serve a precompiled version of this via the web,
and patch it in Javascript as necessary.  We need magic numbers to be
able to find the contents of the structs for search-and-replace.

Signed-off-by: Sean Cross <sean@xobs.io>