support: crt0: fix __dso_handle warning
We were getting a warning when compiling with objects that could
potentially have destructors.

We define this handle so that it exists in the build, even though it is
never used.  It should be a void* pointer, but previously we defined it
as a uint32_t.  These are the same thing on ARM, but it still emitted a
warning.

Fix this difference by re-defining it as a void* pointing to itself.

Signed-off-by: Sean Cross <sean@xobs.io>