sound: sync the current sound engine
Update the sound engine to match the out-of-tree development
version, which is progressing nicely.  The engine is very
much refined, and has a proper bytecode now in addition to a
fully working ADSR for each channel.

The sample song.h demonstrates the use of echo, by having two
channels play the same sound but with a delay on one, and slightly
modified ADSR settings.

It's untested on real hardware, but it ought to work.

Signed-off-by: Sean Cross <sean@xobs.io>